add program to datasets.py module instead of independent script
